app/requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
csscompressor==0.9.5
cssmin==0.2.0
Django==5.0
django-appconf==1.0.6
django-ckeditor==6.7.0
django-compressor==4.4
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-meta==2.4.1
django-resized==1.0.2
django-rosetta==0.10.0
idna==3.6
Jinja2==3.1.2
jsmin==3.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pillow==10.1.0
polib==1.2.0
psycopg==3.1.16
psycopg-binary==3.1.16
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rich==13.7.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uWSGI==2.0.23
